+++ src/types/rx-query.d.ts
@@ -3,34 +3,75 @@ import {
 } from '../rx-query';
 import { StringKeys } from './util';
 
-export interface RxQueryOptions<T> {
-    $eq?: T;
-    $gt?: T;
-    $gte?: T;
-    $lt?: T;
-    $lte?: T;
-    $ne?: T;
-    $in?: T[];
-    $nin?: T[];
-    $regex?: RegExp;
+
+/**
+ * Typed Mango Query Selector
+ * @link https://github.com/mongodb/node-mongodb-native/blob/26bce4a8debb65df5a42dc8599e886c9c83de10d/src/mongo_types.ts
+ * @link https://stackoverflow.com/a/58436959/3443137
+ */
+
+type Join<K, P> = K extends string | number ?
+    P extends string | number ?
+    `${K}${'' extends P ? '' : '.'}${P}`
+    : never : never;
+
+export type Paths<T, D extends number = 10> = [D] extends [never] ? never : T extends object ?
+    { [K in keyof T]-?: K extends string | number ?
+        `${K}` | (Paths<T[K], Prev[D]> extends infer R ? Join<K, R> : never)
+        : never
+    }[keyof T] : '';
+
+export type Leaves<T, D extends number = 10> = [D] extends [never] ? never : T extends object ?
+    { [K in keyof T]-?: Join<K, Leaves<T[K], Prev[D]>> }[keyof T] : '';
+type Prev = [never, 0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10,
+    11, 12, 13, 14, 15, 16, 17, 18, 19, 20, ...0[]];
+
+export type PropertyType<Type, Property extends string> = string extends Property
+    ? unknown
+    : Property extends keyof Type
+    ? Type[Property]
+    : Property extends `${number}`
+    ? Type extends ReadonlyArray<infer ArrayType>
+    ? ArrayType
+    : unknown
+    : Property extends `${infer Key}.${infer Rest}`
+    ? Key extends `${number}`
+    ? Type extends ReadonlyArray<infer ArrayType>
+    ? PropertyType<ArrayType, Rest>
+    : unknown
+    : Key extends keyof Type
+    ? Type[Key] extends Map<string, infer MapType>
+    ? MapType
+    : PropertyType<Type[Key], Rest>
+    : unknown
+    : unknown;
+
+
+export interface MangoQueryOperators<PathValueType> {
+    $eq?: PathValueType;
+    $gt?: PathValueType;
+    $gte?: PathValueType;
+    $lt?: PathValueType;
+    $lte?: PathValueType;
+    $ne?: PathValueType;
+    $in?: PathValueType[];
+    $nin?: PathValueType[];
+    $regex?: string | RegExp;
     $exists?: boolean;
     $type?: 'null' | 'boolean' | 'number' | 'string' | 'array' | 'object';
     $mod?: number;
-    $not?: T;
-    $all?: T[];
+    $not?: PathValueType;
+    $all?: PathValueType[];
     $size?: number;
-    $elemMatch?: RxQueryOptions<T>;
+    $elemMatch?: MangoQuerySelector<PathValueType>;
 }
 
-export type RxQueryObject<T = any> = keyof T & { [P in keyof T]?: T[P] | RxQueryOptions<T[P]>; } & {
-    $or: RxQueryObject<T>[];
-    $nor: RxQueryObject<T>[];
-    $and: RxQueryObject<T>[];
-};
-
-// TODO this should be typed
-export type MangoQuerySelector<RxDocType = any> = {
-    [k: string]: MangoQuerySelector<RxDocType> | any;
+export type MangoQuerySelector<DocType> = Partial<{
+    [Property in Paths<DocType>]: MangoQueryOperators<any> | PropertyType<DocType, Property>;
+}> & {
+    $and?: MangoQuerySelector<DocType>[];
+    $or?: MangoQuerySelector<DocType>[];
+    $nor?: MangoQuerySelector<DocType>[];
 };
 
 /**
@@ -61,7 +102,7 @@ export type MangoQuerySelectorAndIndex<RxDocType = any> = {
     index?: string | string[];
 };
 
-export type MangoQueryNoLimit<RxDocType = any> = MangoQuerySelectorAndIndex<RxDocType> & {
+export type MangoQueryNoLimit<RxDocType> = MangoQuerySelectorAndIndex<RxDocType> & {
     /**
      * Sorting of the results.
      * If no sort is set, RxDB will sort by the primary key.