+++ src/util.ts
@@ -372,7 +372,7 @@ export function removeOneFromArrayIfMatches<T>(ar: T[], condition: (x: T) => boo
     return ar;
 }
 
-function recursiveDeepCopy<T>(o: T | DeepReadonlyObject<T>): T {
+function recursiveDeepCopy<T>(o: T | DeepReadonlyObject<T> | Readonly<T>): T {
     if (!o) return o;
     return deepClone(o, false) as any;
 }
@@ -383,7 +383,7 @@ export const clone = recursiveDeepCopy;
  * is about 3 times faster then using deepClone
  * @link https://jsperf.com/object-rest-spread-vs-clone/2
  */
-export function flatClone<T>(obj: T | DeepReadonlyObject<T>): T {
+export function flatClone<T>(obj: T | DeepReadonlyObject<T> | Readonly<T>): T {
     return Object.assign({}, obj) as any;
 }